+++ autd3capi-gain-holo/src/gspat.rs
@@ -14,9 +14,9 @@ use autd3capi_driver::{
 #[must_use]
 pub unsafe extern "C" fn AUTDGainHoloGSPATSphere(
     backend: BackendPtr,
-    points: *const f64,
-    amps: *const f64,
-    size: u64,
+    points: *const f32,
+    amps: *const f32,
+    size: u32,
     repeat: u32,
     constraint: EmissionConstraintWrap,
 ) -> GainPtr {
@@ -30,9 +30,9 @@ pub unsafe extern "C" fn AUTDGainHoloGSPATSphere(
 #[must_use]
 pub unsafe extern "C" fn AUTDGainHoloGSPATT4010A1(
     backend: BackendPtr,
-    points: *const f64,
-    amps: *const f64,
-    size: u64,
+    points: *const f32,
+    amps: *const f32,
+    size: u32,
     repeat: u32,
     constraint: EmissionConstraintWrap,
 ) -> GainPtr {
