+++ autd3capi-gain-holo/src/sdp.rs
@@ -14,11 +14,11 @@ use autd3capi_driver::{
 #[must_use]
 pub unsafe extern "C" fn AUTDGainHoloSDPSphere(
     backend: BackendPtr,
-    points: *const f64,
-    amps: *const f64,
-    size: u64,
-    alpha: f64,
-    lambda: f64,
+    points: *const f32,
+    amps: *const f32,
+    size: u32,
+    alpha: f32,
+    lambda: f32,
     repeat: u32,
     constraint: EmissionConstraintWrap,
 ) -> GainPtr {
@@ -34,11 +34,11 @@ pub unsafe extern "C" fn AUTDGainHoloSDPSphere(
 #[must_use]
 pub unsafe extern "C" fn AUTDGainHoloSDPT4010A1(
     backend: BackendPtr,
-    points: *const f64,
-    amps: *const f64,
-    size: u64,
-    alpha: f64,
-    lambda: f64,
+    points: *const f32,
+    amps: *const f32,
+    size: u32,
+    alpha: f32,
+    lambda: f32,
     repeat: u32,
     constraint: EmissionConstraintWrap,
 ) -> GainPtr {
