+++ autd3capi/src/geometry/rotation.rs
@@ -4,7 +4,7 @@ use autd3capi_driver::{
 };
 
 #[no_mangle]
-pub unsafe extern "C" fn AUTDRotationFromEulerZYZ(x: f64, y: f64, z: f64, rot: *mut f64) {
+pub unsafe extern "C" fn AUTDRotationFromEulerZYZ(x: f32, y: f32, z: f32, rot: *mut f32) {
     let r = UnitQuaternion::from(EulerAngle::ZYZ(x * rad, y * rad, z * rad));
     rot.add(0).write(r.w);
     rot.add(1).write(r.i);
