+++ autd3capi/src/geometry/transducer.rs
@@ -4,12 +4,12 @@ use autd3capi_driver::*;
 
 #[no_mangle]
 #[must_use]
-pub unsafe extern "C" fn AUTDTransducer(dev: DevicePtr, idx: u32) -> TransducerPtr {
+pub unsafe extern "C" fn AUTDTransducer(dev: DevicePtr, idx: u8) -> TransducerPtr {
     TransducerPtr(&dev[idx as usize] as *const _ as _)
 }
 
 #[no_mangle]
-pub unsafe extern "C" fn AUTDTransducerPosition(tr: TransducerPtr, pos: *mut f64) {
+pub unsafe extern "C" fn AUTDTransducerPosition(tr: TransducerPtr, pos: *mut f32) {
     let p = tr.position();
     pos.add(0).write(p.x);
     pos.add(1).write(p.y);
